requirements.txt
discord.py==2.1.1
openai==0.26.5
python-dotenv==0.21.1
